requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pdfplumber==0.9.0
python-multipart==0.0.6
boto3==1.34.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
